Move branch related stuff into feature folder
